Extract InFlightIndex out of main.cpp
